requirements.txt
Django==4.0.1
psycopg2-binary==2.9.3
djangorestframework==3.13.1
celery==5.1.2
redis==4.1.0
alpha-vantage==2.3.1
gunicorn==20.1.0